Fix lỗi ptimer raccoon nhảy và nếu chết thì ko xài dc, lỗi reset
Raccoon ptimer khi chết lúc flight mode cần reset timer về f8
Lỗi reset do có biến bị xóa mà giữ địa chỉ, là gamefx manager vs effect của nó, sau đó viết lên isOn là 0, trùng hợp là đó cũng là địa chỉ của point 6 pointer nên pointer chỉ sang chỗ khác và corrupt, tuy lỗi hiếm xảy ra nhưng cực kì lắt léo
Các pointer còn lai đều dc xóa và set null đúng 1 lần đảm bảo ko leak memory

+++ PlayScene.h
@@ -63,7 +63,7 @@ public:
 	void SetIsPause();
 
 	bool GetIsStop() { return stopTimer->IsRunning(); }
-	void SetIsStop(float timeSpan) {
+	void SetIsStop(ULONGLONG timeSpan) {
 		stopTimer->SetTimeSpan(timeSpan);
 		stopTimer->Start();
 	}
